Separate Jenkins API and commands in different modules
